requirements.txt
asgiref==3.8.1
cheroot==10.0.1
dj-database-url==2.2.0
Django==5.1
django-appconf==1.0.6
django-debug-toolbar==4.4.6
django-staticfiles==1.2.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
jaraco.functools==4.1.0
more-itertools==10.5.0
mysqlclient==2.2.4
packaging==24.1
pillow==10.4.0
psycopg2==2.9.10
PyJWT==2.9.0
pytz==2024.2
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
whitenoise==6.7.0
